chore: rework listFiles
